% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{barplot}
\alias{barplot}
\title{Bar Plot}
\usage{
barplot(data, xaxisLabel, yaxisLabel, title, flg_darkMode = FALSE)
}
\arguments{
\item{data, }{xAxisValues, yAxisValues,xaxisLabel,yaxisLabel,title,are paramaeters required to plot bar chart. flg_darkMode is boolean variable for dark mode.}
}
\description{
barplot function will plot  a bar chart using echarts4r libary.
}
