% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createPage}
\alias{createPage}
\title{UI function for the Biodiversity App}
\usage{
createPage(title, subtitle, context)
}
\arguments{
\item{title}{Character string for the dashboard title}

\item{subtitle}{Character string for the dashboard subtitle}

\item{context}{Character string to set the app context}
}
\value{
A bs4Dash dashboard page
}
\description{
Creates the main dashboard page structure for the Biodiversity App
}
