% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubsetSelectionModule.R
\name{dataSubsetSelectionServer}
\alias{dataSubsetSelectionServer}
\title{Server Module part for Biodiversity data subset selection}
\usage{
dataSubsetSelectionServer(id, specieOccurrences)
}
\arguments{
\item{id}{Module id}

\item{specieOccurrences}{data.table of the following structure
\code{\link{BiodiversityDashboard::occurrence_pl}}}
}
\value{
reactive \code{data.table}, subset of specieOccurrences input data
}
\description{
Box with a couple of widgets for narrowing down underlying data set
}
\details{
Widgets are:
\itemize{
\item Use scientific name checkbox
\item Specie scientific/vernacular single item selectize. Of those two only one
will be displayed at any given time, which one depends on checkbox state.
Mind that both will always be rendered and updated so that names match.
\item dateRange widget limited to max/min observation dates for selected specie.
Widget will be replaced with 'missing data' info if no specie is selected.
}
}
