% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubsetSelectionModule.R
\name{dataSubsetSelectionUI}
\alias{dataSubsetSelectionUI}
\title{UI Module part for Biodiversity data subset selection}
\usage{
dataSubsetSelectionUI(id)
}
\arguments{
\item{id}{Module id}
}
\description{
UI Module part for Biodiversity data subset selection
}
