% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DistinctChoices}
\alias{DistinctChoices}
\title{Extract unique values from a column based on a filter condition}
\usage{
DistinctChoices(filteredNameID, searchColumnName, selectedColumnName)
}
\arguments{
\item{filteredNameID}{Character. The value to filter by in the search column.}

\item{searchColumnName}{Character. The name of the column to search for the filteredNameID value.}

\item{selectedColumnName}{Character. The name of the column from which to extract unique values.}
}
\value{
A character vector of unique values from the selected column.
}
\description{
This function filters the biodiversity data based on a search value in a specified column,
then extracts unique values from another selected column.
}
