% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{leafletPlot}
\alias{leafletPlot}
\title{Leaflet Plot}
\usage{
leafletPlot(data)
}
\arguments{
\item{data}{used to plot the map}
}
\description{
lealetPlot function will map a selected obeservation by user and distribution of its occurences.
}
