% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils00_load_data.R
\name{load_bd_data}
\alias{load_bd_data}
\title{Load Biodiversity Data}
\usage{
load_bd_data(data = bdData)
}
\arguments{
\item{data}{A data frame. The biodiversity dataset to filter.}
}
\value{
A tibble containing the processed biodiversity data or NULL if an error occurs
}
\description{
This function loads internally stored biodiversity data,
selects variables required for this project, and returns the processed data.
It handles null data with appropriate error messaging.
}
