% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils00_loadData.R
\name{load_bdData}
\alias{load_bdData}
\title{Load Biodiversiry Data}
\usage{
load_bdData()
}
\description{
load_bdData function  will use internal stored data select veritable required for this project and return it.
}
