% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod01_loadInputs.R
\name{mod01_loadInputs_ui}
\alias{mod01_loadInputs_ui}
\title{loadInputs UI Function}
\usage{
mod01_loadInputs_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
mod01_loadInputs is first module which creates radio button for species selection based on scientific names and vernacular names, then based on user choice create dynamic searchable dropbox for species scienific and vernacular name.
}
