% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadData.R
\name{mod_loadData_server}
\alias{mod_loadData_server}
\title{loadData Server Functions}
\usage{
mod_loadData_server(id, searchSpeciesIds = NULL)
}
\arguments{
\item{searchSpeciesIds}{is a reactive list which contains Scientific Names and Vernacular Name of species selected by user}
}
\description{
loadData Server Functions
}
