% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadData.R
\name{mod_loadData_ui}
\alias{mod_loadData_ui}
\title{loadData UI Function}
\usage{
mod_loadData_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
mod_loadData will load the give user to choose/search species by there scientific names or vernacular name
}
