% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{monthlyOccurence}
\alias{monthlyOccurence}
\title{Monthly Occurence of Animals}
\usage{
monthlyOccurence(data)
}
\arguments{
\item{data}{is a data.frame which contain date format column, datecolumn is a date format column}
}
\description{
monthlyOccurence will count the occurence per month in chronological order
}
