% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_bd_app}
\alias{run_bd_app}
\title{Run Biodiversity Dashboard}
\usage{
run_bd_app(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{shinyApp}}
}
\value{
A Shiny application object
}
\description{
This function launches the Biodiversity Dashboard Shiny application.
}
\examples{
\dontrun{
run_bd_app()
}
}
