% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieOccurrencesMapModule.R
\name{specieOccurrencesMapServer}
\alias{specieOccurrencesMapServer}
\title{Server Module part for species occurrence map display}
\usage{
specieOccurrencesMapServer(id, speciesOccurrences, specieOccurrences)
}
\arguments{
\item{id}{Module id}

\item{speciesOccurrences}{\code{data.frame} containing full data,
\code{\link{BiodiversityDashboard::occurrence_pl}}}

\item{specieOccurrences}{\code{data.frame} containing filtered data,
as returned by \code{\link{BiodiversityDashboard::dataSubsetSelectionServer()}}}
}
\value{
named list with currently displayed map boundaries
}
\description{
Leaflet map with specie occurrences displayed as dots
}
\details{
Map is limited to are covered by the dataset. Dots size scales with observed
individual count. Count is a sum of all observation at that specific point
across whole dataset. If dataset is empty prompt to select proper subset is
displayed. Map zoom in level is preserved across specie/date range changes.
}
