% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieOccurrencesMapModule.R
\name{specieOccurrencesMapUI}
\alias{specieOccurrencesMapUI}
\title{UI Module part for species occurrence map display}
\usage{
specieOccurrencesMapUI(id)
}
\arguments{
\item{id}{Module id}
}
\description{
UI Module part for species occurrence map display
}
