% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieOccurrencesTimelinePlotModule.R
\name{specieOccurrencesTimelinePlotServer}
\alias{specieOccurrencesTimelinePlotServer}
\title{Server Module part for species occurrence timeline plot}
\usage{
specieOccurrencesTimelinePlotServer(id, subsetSelected, displayedMapBoundaries)
}
\arguments{
\item{id}{Module id}

\item{subsetSelected}{\code{data.frame} containing filtered data,
as returned by \code{\link{BiodiversityDashboard::dataSubsetSelectionServer()}}}

\item{displayedMapBoundaries}{named list with coordinates span to whin plot data will be limited to.
By default taken as return from \code{\link{BiodiversityDashboard::specieOccurrencesMapServer()}}}
}
\description{
Bar plot displaying observed specie individuals over time
}
\details{
Bar plot is accompanies by checkbot. When ticked it limits plot data to
coordinates span provided via argument \code{displayedMapBoundaries}. Is
intended to be used with \code{\link{BiodiversityDashboard::specieOccurrencesMapServer()}}
with aforementioned coordinates being modules output.
}
