% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieOccurrencesTimelinePlotModule.R
\name{specieOccurrencesTimelinePlotUI}
\alias{specieOccurrencesTimelinePlotUI}
\title{UI Module part for specie occurrences timeline plot}
\usage{
specieOccurrencesTimelinePlotUI(id)
}
\arguments{
\item{id}{Module id}
}
\description{
UI Module part for specie occurrences timeline plot
}
