% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateDropdownBasedOnRadioBtnValues}
\alias{updateDropdownBasedOnRadioBtnValues}
\title{Update dropdown inputs based on radio button selection}
\usage{
updateDropdownBasedOnRadioBtnValues(
  radioBtn_search_byName,
  radioBtn_search_byNameID_ver,
  id_ver_name,
  scientificName_str,
  vernacularName_str,
  id_selected_sci_name,
  radioBtn_search_byNameID_sci,
  id_sci_name,
  id_selected_ver_name,
  session
)
}
\arguments{
\item{radioBtn_search_byName}{The current value of the radio button that determines
the search method (vernacular or scientific name)}

\item{radioBtn_search_byNameID_ver}{The value of the radio button that corresponds to
searching by vernacular name}

\item{id_ver_name}{The input value from the vernacular name dropdown}

\item{scientificName_str}{Constant string identifier for scientific name, typically "scientificName"}

\item{vernacularName_str}{Constant string identifier for vernacular name, typically "vernacularName"}

\item{id_selected_sci_name}{The input value from the matching scientific name dropdown}

\item{radioBtn_search_byNameID_sci}{The value of the radio button that corresponds to
searching by scientific name}

\item{id_sci_name}{The input value from the scientific name dropdown}

\item{id_selected_ver_name}{The input value from the matching vernacular name dropdown}

\item{session}{The Shiny session object}
}
\description{
Updates searchable dropdown inputs for species selection based on the user's choice
of searching by vernacular name or scientific name. When a user selects a species
in one dropdown, this function updates the corresponding value in the other dropdown.
}
