% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{updateSpecies}
\alias{updateSpecies}
\title{Update Species Dropdown}
\usage{
updateSpecies(session, id, choices)
}
\arguments{
\item{session, }{id, choices used to give input session, input id, and list of all choices for selecting species byVernacular name and by scientific name}
}
\description{
updateSpecies will update a searchbox and dropdown for selecting species byVernacular name and by scientific name in the server
}
