% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createPage}
\alias{createPage}
\title{This file will create helper functions required for structure buiding}
\usage{
createPage(title, subtitle, context)
}
\arguments{
\item{title, }{subtitle, context for header , exra informaton and to set context respectivly.}
}
\description{
This file will create helper functions required for structure buiding
}
