% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DistinctChoices}
\alias{DistinctChoices}
\title{Unique Species}
\usage{
DistinctChoices(filteredNameID, searchColumnName, selectedColumnName)
}
\arguments{
\item{filteredNameID}{is Search Species by Vernacular Name or Scientific Name,
selectedColumnName is selected column, SearchColumnName colum from user is seaching the species.}
}
\description{
DistinctChoices will Choose species from scientific name species given Vernacular Name species, distinct() select unique values and pull() convert data.frame to vector
}
