% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dropdownBasedOnRadioBtn}
\alias{dropdownBasedOnRadioBtn}
\title{Create Species Dropdown}
\usage{
dropdownBasedOnRadioBtn(
  radioBtn_search_byName,
  radioBtn_search_byNameID,
  id1,
  label1,
  choices1,
  id2,
  label2,
  choices2
)
}
\arguments{
\item{radioBtn_search_byName}{andradioBtn_search_byNameID are Search by Name radio button input value and specific choice like Vernacular Name or Scientific Name.
id1, id2, label1, label2, choices1,choices2  used to give input ids, labels and list of all choices for selecting species byVernacular name and by scientific name}
}
\description{
dropdownBasedOnRadioBtn will create a searchable dropdown for selecting species byVernacular name and by scientific name based on radio button for Search species by Vernacuar Name or Scientific Name
}
