% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod01_loadInputs.R
\name{mod01_loadInputs_ui}
\alias{mod01_loadInputs_ui}
\title{Load Inputs UI Function}
\usage{
mod01_loadInputs_ui(id)
}
\arguments{
\item{id}{A character string used to identify the namespace for the module}
}
\value{
A Shiny UI element
}
\description{
This module creates input controls for species selection. It displays a radio button
for choosing between scientific names and vernacular names, then based on the selection
creates a dynamic searchable dropdown for the corresponding species names.
}
