% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SelectedbdData}
\alias{SelectedbdData}
\title{User Extracted Data}
\usage{
SelectedbdData(
  radioBtn_search_byName,
  radioBtn_search_byNameID,
  searchByVerOrSciName,
  selectedByVerOrSciName
)
}
\arguments{
\item{searchedVerncularNameID}{and searchedScientifNameID are input ids of Search Species by Vernacular Name and Search Species by scientific Name}
}
\description{
SelectedbdData is subset of biodiversity data corresponding to search by user search for species by their vernacularName and scientificName.
}
