% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateSpecies}
\alias{updateSpecies}
\title{Update Species Dropdown}
\usage{
updateSpecies(session, id, choices)
}
\arguments{
\item{session}{The Shiny session object}

\item{id}{The input ID of the selectize input to update}

\item{choices}{A character vector of available species choices to display in the dropdown}
}
\value{
No return value; called for side effects
}
\description{
updateSpecies will update a searchbox and dropdown for selecting species byVernacular name and by scientific name in the server
}
